server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
feedparser==6.0.10
jinja2==3.1.3
playwright==1.41.0
pillow==10.2.0
python-dotenv==1.0.0
